requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
openai>=1.0.0
groq>=0.4.0
requests>=2.25.0
unidic-lite
typing-extensions>=4.0.0
transformers>=4.20.0
torch>=1.12.0
sentencepiece>=0.1.95
fugashi>=1.1.0
ipadic>=1.0.0
aiohttp>=3.8.0
aiofiles>=23.0.0
asyncio-throttle>=1.0.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0